requirements.txt
# Core ML/AI libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.20.0
peft>=0.5.0
bitsandbytes>=0.41.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# YouTube API and web scraping
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1
beautifulsoup4>=4.12.0
requests>=2.31.0

# NLP and text processing
nltk>=3.8.1
spacy>=3.7.0
textstat>=0.7.3

# Web interface
gradio>=3.45.0
streamlit>=1.25.0
fastapi>=0.103.0
uvicorn>=0.23.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.7

# Development tools
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
pre-commit>=3.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.14.0